#Observação da interação com as CheckBox e salvando no BD

+++ app/src/main/java/com/example/tellcom/service/model/ScoreModel.kt
@@ -11,7 +11,7 @@ data class ScoreModel (
     val jobName: String,
     val singlePoints: Double,
     val metaPoints: Double,
-    val currentScore: Int = 0
+    val currentScore: Double = 0.0
 )
